+++ common/src/main/java/com/verdantartifice/primalmagick/common/tiles/crafting/RunecarvingTableTileEntity.java
@@ -16,6 +16,7 @@ import net.minecraft.world.entity.player.Player;
 import net.minecraft.world.inventory.AbstractContainerMenu;
 import net.minecraft.world.item.Items;
 import net.minecraft.world.level.block.state.BlockState;
+import org.jetbrains.annotations.NotNull;
 
 import java.util.Optional;
 
@@ -56,7 +57,7 @@ public abstract class RunecarvingTableTileEntity extends AbstractTileSidedInvent
     }
 
     @Override
-    public Optional<Integer> getInventoryIndexForFace(Direction face) {
+    public Optional<Integer> getInventoryIndexForFace(@NotNull Direction face) {
         return switch (face) {
             case DOWN -> Optional.empty();
             default -> Optional.of(INPUT_INV_INDEX);