+++ common/src/main/java/com/verdantartifice/primalmagick/common/tiles/devices/ScribeTableTileEntity.java
@@ -27,6 +27,7 @@ import net.minecraft.world.item.ItemStack;
 import net.minecraft.world.item.Items;
 import net.minecraft.world.level.Level;
 import net.minecraft.world.level.block.state.BlockState;
+import org.jetbrains.annotations.NotNull;
 
 import java.util.Optional;
 
@@ -80,7 +81,7 @@ public abstract class ScribeTableTileEntity extends AbstractTileSidedInventoryPM
     }
 
     @Override
-    public Optional<Integer> getInventoryIndexForFace(Direction face) {
+    public Optional<Integer> getInventoryIndexForFace(@NotNull Direction face) {
         return switch (face) {
             case DOWN -> Optional.of(OUTPUT_INV_INDEX);
             default -> Optional.of(INPUT_INV_INDEX);