+++ common/src/main/java/com/verdantartifice/primalmagick/common/tiles/mana/AutoChargerTileEntity.java
@@ -18,6 +18,7 @@ import net.minecraft.world.item.ItemStack;
 import net.minecraft.world.level.Level;
 import net.minecraft.world.level.block.state.BlockState;
 import net.minecraft.world.phys.Vec3;
+import org.jetbrains.annotations.NotNull;
 
 import java.util.HashSet;
 import java.util.Optional;
@@ -136,7 +137,7 @@ public abstract class AutoChargerTileEntity extends AbstractTileSidedInventoryPM
     }
 
     @Override
-    public Optional<Integer> getInventoryIndexForFace(Direction face) {
+    public Optional<Integer> getInventoryIndexForFace(@NotNull Direction face) {
         return Optional.of(INPUT_INV_INDEX);
     }