+++ common/src/main/java/com/verdantartifice/primalmagick/common/tiles/misc/CarvedBookshelfTileEntity.java
@@ -17,6 +17,7 @@ import net.minecraft.world.level.block.state.BlockState;
 import net.minecraft.world.level.gameevent.GameEvent;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
+import org.jetbrains.annotations.NotNull;
 
 import java.util.Objects;
 import java.util.Optional;
@@ -57,7 +58,7 @@ public abstract class CarvedBookshelfTileEntity extends AbstractTileSidedInvento
     }
 
     @Override
-    public Optional<Integer> getInventoryIndexForFace(Direction face) {
+    public Optional<Integer> getInventoryIndexForFace(@NotNull Direction face) {
         return Optional.of(INPUT_INV_INDEX);
     }